Initial refactor for review
