+++ src/Cannon/src/com/frc2508/cannon/Main.java
@@ -6,14 +6,9 @@
 /*----------------------------------------------------------------------------*/
 package com.frc2508.cannon;
 
-import edu.wpi.first.wpilibj.CANJaguar;
 import edu.wpi.first.wpilibj.DigitalInput;
-import edu.wpi.first.wpilibj.GenericHID;
 import edu.wpi.first.wpilibj.Relay;
-import edu.wpi.first.wpilibj.RobotDrive;
 import edu.wpi.first.wpilibj.SimpleRobot;
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj.can.CANTimeoutException;
 
 /**
  * The VM is configured to automatically run this class, and to call the
@@ -30,10 +25,7 @@ public class Main extends SimpleRobot {
     public void autonomous() {
 
     }
-    
-    /*Tunables*/
-    double deadBand = .05;
-    
+       
     /*Initialization*/
     Relay firingSolenoidRelay = new Relay(1, Relay.Direction.kForward);
     Relay tiltRelay = new Relay(2, Relay.Direction.kBoth);
@@ -41,46 +33,15 @@ public class Main extends SimpleRobot {
     Relay compressor1Relay = new Relay(4, Relay.Direction.kReverse);
     DigitalInput pressureSwitch = new DigitalInput(1);
 
-    CANJaguar[] driveMotors = new CANJaguar[4];
-
-    RobotDrive robotDrive;
+    Drive robotDrive;
 
     Gamepad gamepad = new Gamepad(1);
-
-    private void configJaguar(int index, int id) {
-        try {
-            CANJaguar jag = new CANJaguar(id, CANJaguar.ControlMode.kSpeed);
-            jag.setSpeedReference(CANJaguar.SpeedReference.kQuadEncoder);
-            jag.setSafetyEnabled(false);
-            //jag.configMaxOutputVoltage(100);
-            jag.configEncoderCodesPerRev(360);
-            jag.setPID(.12, .02, .05);
-            jag.enableControl();
-
-            driveMotors[index] = jag;
-        } catch (CANTimeoutException ex) {
-            ex.printStackTrace();
-        }
-    }
-
-    public void configureDrive() {
-        configJaguar(0, 8);
-        configJaguar(1, 2);
-        configJaguar(2, 3);
-        configJaguar(3, 4);
-        robotDrive = new RobotDrive(driveMotors[1], driveMotors[0], driveMotors[2], driveMotors[3]);
-        robotDrive.setInvertedMotor(RobotDrive.MotorType.kRearLeft, true);
-
-        double gearRatio = 12.0;
-        double maxRPM = 5310.0;
-        robotDrive.setMaxOutput((maxRPM / gearRatio) / 1.5);
-    }
 
     /*
      * This function is called once each time the robot enters operator control.
      */
     public void operatorControl() {
-        configureDrive();
+        robotDrive = new Drive();
         long last = System.currentTimeMillis();
         while (isOperatorControl() && isEnabled()) {
             long millisPerTick = System.currentTimeMillis() - last;
@@ -91,12 +52,13 @@ public class Main extends SimpleRobot {
             checkPressure(millisPerTick);
             doFire(millisPerTick);
             doTilt(millisPerTick);
-            doArcade(millisPerTick);
+            robotDrive.DoArcade(millisPerTick, gamepad.getRawAxis(1), gamepad.getRawAxis(2));
             //printMoveStickAxis();
 
             //Timer.delay(.01);
         }
-        rampPairThing = new Pair(0,0);
+        
+        robotDrive.ResetRampPair();        
     }
 
     private long compressorDelay = 0;
@@ -161,95 +123,12 @@ public class Main extends SimpleRobot {
         //System.out.println("Firing relay state: " + firingSolenoidRelay.get().value);        
     }
 
-    public static double ramp(double input) {
-        if (input == 0) {
-            return 0;
-        }
-
-        return input * Math.abs(input);
-    }
-
-    public double deadband(double input) {        
-        return input > deadBand || input < -1 * deadBand ? input : 0;
-    }    
-    	
-    public Pair squareTheCircle(Pair input)
-    {
-        double shortSide = Math.min(Math.abs(input.x),Math.abs(input.y));
-        double longSide = Math.max(Math.abs(input.x),Math.abs(input.y));
-
-        double r = Math.sqrt((shortSide*shortSide)+(longSide*longSide));
-
-        double scaleFactor = (longSide/r);
-
-        return new Pair(input.x/scaleFactor,input.y/scaleFactor);
-    }
-
     /**
      * This function is called once each time the robot enters test mode.
      */
     public void test() {
 
     }    	
-
-    private void doMecanum(long millisPerTick) {
-        double inX = gamepad.getX();
-        double inY = gamepad.getY();
-        double inTwist = gamepad.getTwist();
-
-        double outX = deadband(inX);
-        double outY = deadband(inY);
-
-        double invertedOutY = -1 * outY;
-        double invertedTwist = -1 * inTwist;
-
-        outX = ramp(outX);
-        // outY = ramp(outY);
-        invertedOutY = ramp(invertedOutY);
-        invertedTwist = ramp(invertedTwist);
-
-        robotDrive.mecanumDrive_Cartesian(outX, invertedOutY, invertedTwist, 0);
-    }
-
-    private void doTank(long millisPerTick) {
-        double inLeft = gamepad.getRawAxis(2);
-        double inRight = -gamepad.getRawAxis(4);
-
-        double outLeft = ramp(deadband(inLeft));
-        double outRight = ramp(deadband(inRight));
-
-        robotDrive.tankDrive(outLeft, outRight);
-    }
-
-    private void doArcade(long millisPerTick) {
-        double inRotate = gamepad.getRawAxis(1);
-        inRotate = -1 * inRotate;
-        
-        double inMove = gamepad.getRawAxis(2);
-        inMove = -1 * inMove;
-        
-        Pair pair = squareTheCircle(doTimeRamp(millisPerTick,new Pair(inRotate, inMove)));
-        
-        robotDrive.arcadeDrive(deadband(pair.x), deadband(pair.y), true);
-    }
-    
-    Pair rampPairThing = new Pair(0,0);
-    double slewRate = .10;
-    
-    private Pair doTimeRamp(long millisPerTick, Pair pair)
-    {
-        double posSlewLimit = slewRate*1000/millisPerTick;
-        double negSlewLimit = -posSlewLimit;
-        double xDiff = pair.x - rampPairThing.x;
-        double yDiff = pair.y - rampPairThing.y;
-       
-        xDiff = Math.max(negSlewLimit, Math.min(posSlewLimit, xDiff));
-        yDiff = Math.max(negSlewLimit, Math.min(posSlewLimit, yDiff));
-        
-        rampPairThing = new Pair(rampPairThing.x + xDiff, rampPairThing.y + yDiff);
-        
-        return rampPairThing;
-    }
 
     private void printMoveStickAxis() {
         for (int i = 0; i < 6; i++) {
